requirements.txt
Babel==2.13.1
beautifulsoup4==4.12.2
ipykernel==6.27.1
numpy==1.26.2
pandas==2.1.3
pdf2image==1.16.3
PyPDF2==3.0.1
pytesseract==0.3.10
replicate==0.21.0
requests==2.31.0
streamlit==1.28.2
streamlit-jupyter==0.2.1
google-generativeai
easyocr==1.7.1
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
pypdfium==0.0.15
pypdfium2==4.27.0
